+++ index.html
@@ -79,9 +79,7 @@
             </div>
         </div>
 
-
         
-
 
         <script type="module"src="js/loadComponents.js"></script>
         <!-- <script type="module" src="js/middleCalendar.js"></script> -->
