+++ js/listener/paymentMethod.js
@@ -1,0 +1,3 @@
+export function initPaymentMethodListener() {
+    
+}
