+++ js/state/data.js
@@ -1,0 +1,13 @@
+export const category = {
+    "생활": "--colorchip-90",
+    "식비": "--colorchip-60",
+    "교통": "--colorchip-70",
+    "쇼핑/뷰티": "--colorchip-30",
+    "의료/건강": "--colorchip-50",
+    "문화/여가": "--colorchip-100",
+    미분류: "--colorchip-110",
+    월급: "--colorchip-20",
+    용돈: "--colorchip-40",
+    기타수입: "--colorchip-10"
+}
+
